a2a_1_adk_client_and_remote/requirements.txt
google-search-results
python-dotenv
google-adk
google-adk[a2a]
uvicorn